extract 2010 import
